dnc-300-movie-finder-dataTests/Controllers/MovieFinderControllerTests.cs
using Microsoft.VisualStudio.TestTools.UnitTesting;
using System;
using System.Collections.Generi,sc;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace dnc_300_movie_finder_data.Controllers.Tests
{
    [TestClass]
    public class MovieFinderControllerTests
    {
        [TestMethod]
        public void TestFindMovieByTitle()
        {
            MovieFinderController mfc = new MovieFinderController();
            string response = mfc.FindMovie("t", "Batman");
            Assert.IsTrue(response.Contains("Batman"));
        }
        [TestMethod]
        public void TestFindMovieByTitleWithCache()
        {
            MovieFinderController mfc = new MovieFinderController();
            string response = mfc.FindMovie("t", "baby%20driver");
            response = mfc.FindMovie(t: "baby%20driver");
            Assert.IsTrue(response.Contains("Baby Driver"));
        }
        [TestMethod]
        public void TestFindMovieByID()
        {
            MovieFinderController mfc = new MovieFinderController();
            string response = mfc.FindMovie("i", "tt3896198");
            Assert.IsTrue(response.Contains("Guardians of the Galaxy"));
        }
        [TestMethod]
        public void TestFindMovieByIDWithCache()
        {
            MovieFinderController mfc = new MovieFinderController();
            string response = mfc.FindMovie("i", "tt3896198 ");
            response = mfc.FindMovie("i", "tt3896198 ");
            Assert.IsTrue(response.Contains("Guardians of the Galaxy"));
        }
    }
}